+++ internal/service/operations/tags_test.go
@@ -5,8 +5,10 @@ import (
 	"os"
 	"os/exec"
 	"path"
+	"strings"
 	"testing"
 
+	"gitlab.com/gitlab-org/gitaly/internal/git/log"
 	"gitlab.com/gitlab-org/gitaly/internal/testhelper"
 
 	pb "gitlab.com/gitlab-org/gitaly-proto/go"
@@ -91,7 +93,126 @@ func TestSuccessfulGitHooksForUserDeleteTagRequest(t *testing.T) {
 	}
 }
 
-func TestFailedUserDeleteTagDueToValidation(t *testing.T) {
+func TestSuccessfulUserCreateTagRequest(t *testing.T) {
+	ctx, cancel := testhelper.Context()
+	defer cancel()
+
+	server := runOperationServiceServer(t)
+	defer server.Stop()
+
+	client, conn := newOperationClient(t)
+	defer conn.Close()
+
+	targetRevision := "c7fbe50c7c7419d9701eebe64b1fdacc3df5b9dd"
+	targetRevisionCommit, err := log.GetCommit(ctx, testRepo, targetRevision, "")
+	require.NoError(t, err)
+
+	user := &pb.User{
+		Name:  []byte("Ahmad Sherif"),
+		Email: []byte("ahmad@gitlab.com"),
+		GlId:  "user-123",
+	}
+	inputTagName := "to-be-created-soon"
+
+	testCases := []struct {
+		desc           string
+		tagName        string
+		message        string
+		targetRevision string
+		expectedTag    *pb.Tag
+	}{
+		{
+			desc:           "lightweight tag",
+			tagName:        inputTagName,
+			targetRevision: targetRevision,
+			expectedTag: &pb.Tag{
+				Name:         []byte(inputTagName),
+				TargetCommit: targetRevisionCommit,
+			},
+		},
+		{
+			desc:           "annotated tag",
+			tagName:        inputTagName,
+			targetRevision: targetRevision,
+			message:        "This is an annotated tag",
+			expectedTag: &pb.Tag{
+				Name:         []byte(inputTagName),
+				TargetCommit: targetRevisionCommit,
+				Message:      []byte("This is an annotated tag"),
+			},
+		},
+	}
+
+	for _, testCase := range testCases {
+		t.Run(testCase.desc, func(t *testing.T) {
+			request := &pb.UserCreateTagRequest{
+				Repository:     testRepo,
+				TagName:        []byte(inputTagName),
+				TargetRevision: []byte(testCase.targetRevision),
+				User:           user,
+				Message:        []byte(testCase.message),
+			}
+
+			ctx, cancel := testhelper.Context()
+			defer cancel()
+
+			response, err := client.UserCreateTag(ctx, request)
+			defer exec.Command("git", "-C", testRepoPath, "tag", "-d", inputTagName).Run()
+
+			id := testhelper.MustRunCommand(t, nil, "git", "-C", testRepoPath, "rev-parse", inputTagName)
+			testCase.expectedTag.Id = strings.TrimSpace(string(id))
+
+			require.NoError(t, err)
+			require.Equal(t, testCase.expectedTag, response.Tag)
+			require.Empty(t, response.PreReceiveError)
+
+			tag := testhelper.MustRunCommand(t, nil, "git", "-C", testRepoPath, "tag")
+			require.Contains(t, string(tag), inputTagName)
+		})
+	}
+}
+
+func TestSuccessfulGitHooksForUserCreateTagRequest(t *testing.T) {
+	server := runOperationServiceServer(t)
+	defer server.Stop()
+
+	client, conn := newOperationClient(t)
+	defer conn.Close()
+
+	tagName := "new-tag"
+	user := &pb.User{
+		Name:  []byte("Ahmad Sherif"),
+		Email: []byte("ahmad@gitlab.com"),
+		GlId:  "user-123",
+	}
+	request := &pb.UserCreateTagRequest{
+		Repository:     testRepo,
+		TagName:        []byte(tagName),
+		TargetRevision: []byte("c7fbe50c7c7419d9701eebe64b1fdacc3df5b9dd"),
+		User:           user,
+	}
+
+	for _, hookName := range []string{"pre-receive", "update", "post-receive"} {
+		t.Run(hookName, func(t *testing.T) {
+			defer exec.Command("git", "-C", testRepoPath, "tag", "-d", tagName).Run()
+
+			hookPath, hookOutputTempPath := writeEnvToHook(t, hookName)
+			defer os.Remove(hookPath)
+
+			ctx, cancel := testhelper.Context()
+			defer cancel()
+
+			response, err := client.UserCreateTag(ctx, request)
+			require.NoError(t, err)
+			require.Empty(t, response.PreReceiveError)
+
+			output := string(testhelper.MustReadFile(t, hookOutputTempPath))
+			require.Contains(t, output, "GL_ID="+user.GlId)
+		})
+	}
+}
+
+func TestFailedUserDeleteTagRequestDueToValidation(t *testing.T) {
 	server := runOperationServiceServer(t)
 	defer server.Stop()
 
@@ -186,6 +307,140 @@ func TestFailedUserDeleteTagDueToHooks(t *testing.T) {
 
 			tags := testhelper.MustRunCommand(t, nil, "git", "-C", testRepoPath, "tag")
 			require.Contains(t, string(tags), tagNameInput, "tag name does not exist in tags list")
+		})
+	}
+}
+
+func TestFailedUserCreateTagDueToHooks(t *testing.T) {
+	server := runOperationServiceServer(t)
+	defer server.Stop()
+
+	client, conn := newOperationClient(t)
+	defer conn.Close()
+
+	user := &pb.User{
+		Name:  []byte("Ahmad Sherif"),
+		Email: []byte("ahmad@gitlab.com"),
+		GlId:  "user-123",
+	}
+	request := &pb.UserCreateTagRequest{
+		Repository:     testRepo,
+		TagName:        []byte("new-tag"),
+		TargetRevision: []byte("c7fbe50c7c7419d9701eebe64b1fdacc3df5b9dd"),
+		User:           user,
+	}
+
+	hookContent := []byte("#!/bin/sh\necho GL_ID=$GL_ID\nexit 1")
+
+	for _, hookName := range []string{"pre-receive", "update"} {
+		hookPath := path.Join(testRepoPath, "hooks", hookName)
+		ioutil.WriteFile(hookPath, hookContent, 0755)
+		defer os.Remove(hookPath)
+
+		ctx, cancel := testhelper.Context()
+		defer cancel()
+
+		response, err := client.UserCreateTag(ctx, request)
+		require.Nil(t, err)
+		require.Contains(t, response.PreReceiveError, "GL_ID="+user.GlId)
+	}
+}
+
+func TestFailedUserCreateTagRequestDueToTagExistence(t *testing.T) {
+	server := runOperationServiceServer(t)
+	defer server.Stop()
+
+	client, conn := newOperationClient(t)
+	defer conn.Close()
+
+	user := &pb.User{
+		Name:  []byte("Ahmad Sherif"),
+		Email: []byte("ahmad@gitlab.com"),
+		GlId:  "user-123",
+	}
+
+	testCase := struct {
+		tagName        string
+		targetRevision string
+		user           *pb.User
+	}{
+		tagName:        "v1.1.0",
+		targetRevision: "master",
+		user:           user,
+	}
+
+	request := &pb.UserCreateTagRequest{
+		Repository:     testRepo,
+		TagName:        []byte(testCase.tagName),
+		TargetRevision: []byte(testCase.targetRevision),
+		User:           testCase.user,
+	}
+
+	ctx, cancel := testhelper.Context()
+	defer cancel()
+
+	response, err := client.UserCreateTag(ctx, request)
+	require.NoError(t, err)
+	require.Equal(t, response.Exists, true)
+}
+
+func TestFailedUserCreateTagRequestDueToValidation(t *testing.T) {
+	server := runOperationServiceServer(t)
+	defer server.Stop()
+
+	client, conn := newOperationClient(t)
+	defer conn.Close()
+
+	user := &pb.User{
+		Name:  []byte("Ahmad Sherif"),
+		Email: []byte("ahmad@gitlab.com"),
+		GlId:  "user-123",
+	}
+
+	testCases := []struct {
+		desc           string
+		tagName        string
+		targetRevision string
+		user           *pb.User
+		code           codes.Code
+	}{
+		{
+			desc:           "empty target revision",
+			tagName:        "shiny-new-tag",
+			targetRevision: "",
+			user:           user,
+			code:           codes.InvalidArgument,
+		},
+		{
+			desc:           "empty user",
+			tagName:        "shiny-new-tag",
+			targetRevision: "master",
+			user:           nil,
+			code:           codes.InvalidArgument,
+		},
+		{
+			desc:           "non-existing starting point",
+			tagName:        "new-tag",
+			targetRevision: "i-dont-exist",
+			user:           user,
+			code:           codes.FailedPrecondition,
+		},
+	}
+
+	for _, testCase := range testCases {
+		t.Run(testCase.desc, func(t *testing.T) {
+			request := &pb.UserCreateTagRequest{
+				Repository:     testRepo,
+				TagName:        []byte(testCase.tagName),
+				TargetRevision: []byte(testCase.targetRevision),
+				User:           testCase.user,
+			}
+
+			ctx, cancel := testhelper.Context()
+			defer cancel()
+
+			_, err := client.UserCreateTag(ctx, request)
+			testhelper.AssertGrpcError(t, err, testCase.code, "")
 		})
 	}
 }