+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Gitaly changelog
 
+UNRELEASED
+
+- Implement UserCreateTag RPC
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/374
+
 v0.41.0
 
 - Wait for monitor goroutine to return during supervisor shutdown